+++ drools-platform-entity/src/main/java/org/chtijbug/drools/platform/entity/event/PlatformKnowledgeBaseInitialConnectionEvent.java
@@ -31,7 +31,7 @@ public class PlatformKnowledgeBaseInitialConnectionEvent extends KnowledgeBaseEv
     private int port;
     private Date startDate;
 
-    public PlatformKnowledgeBaseInitialConnectionEvent(int eventID, Date dateEvent, int ruleBaseID) {
+    public PlatformKnowledgeBaseInitialConnectionEvent(Long eventID, Date dateEvent, Long ruleBaseID) {
         super(eventID, dateEvent, ruleBaseID);
         this.startDate = dateEvent;
     }
