+++ drools-platform-manager/drools-platform-persistence/src/main/java/org/chtijbug/drools/platform/persistence/PlatformRuntimeDefinitionRepository.java
@@ -31,7 +31,7 @@ import java.util.List;
 public interface PlatformRuntimeDefinitionRepository extends JpaRepository<PlatformRuntimeDefinition, Long> {
 
 
-    PlatformRuntimeDefinition findByRuleBaseID(Integer ruleBaseId);
+    PlatformRuntimeDefinition findByRuleBaseID(Long ruleBaseId);
 
     List<PlatformRuntimeDefinition> findAll();
