+++ drools-platform-manager/drools-platform-persistence/src/main/java/org/chtijbug/drools/platform/persistence/pojo/FireAllRulesExecution.java
@@ -30,15 +30,15 @@ public class FireAllRulesExecution {
     private Long executionTime;
 
 
-    private Integer startEventID;
+    private Long startEventID;
 
 
     private FireAllRulesExecutionStatus fireAllRulesExecutionStatus;
 
 
-    private Integer stopEventID;
+    private Long stopEventID;
 
-    private Integer maxRulesEventID;
+    private Long maxRulesEventID;
 
     public FireAllRulesExecution() {
     }
@@ -61,11 +61,11 @@ public class FireAllRulesExecution {
         this.endDate = endDate;
     }
 
-    public Integer getStartEventID() {
+    public Long getStartEventID() {
         return startEventID;
     }
 
-    public void setStartEventID(Integer startEventID) {
+    public void setStartEventID(Long startEventID) {
         this.startEventID = startEventID;
     }
 
@@ -104,19 +104,19 @@ public class FireAllRulesExecution {
     }
 
 
-    public Integer getStopEventID() {
+    public Long getStopEventID() {
         return stopEventID;
     }
 
-    public void setStopEventID(Integer stopEventID) {
+    public void setStopEventID(Long stopEventID) {
         this.stopEventID = stopEventID;
     }
 
-    public Integer getMaxRulesEventID() {
+    public Long getMaxRulesEventID() {
         return maxRulesEventID;
     }
 
-    public void setMaxRulesEventID(Integer maxRulesEventID) {
+    public void setMaxRulesEventID(Long maxRulesEventID) {
         this.maxRulesEventID = maxRulesEventID;
     }
