+++ drools-platform-manager/drools-platform-persistence/src/main/java/org/chtijbug/drools/platform/persistence/pojo/RuleExecution.java
@@ -29,9 +29,9 @@ public class RuleExecution {
     private Date startDate;
     private Date endDate;
 
-    private Integer startEventID;
+    private Long startEventID;
 
-    private Integer stopEventID;
+    private Long stopEventID;
 
 
     private List<Fact> whenFacts = new ArrayList<Fact>();
@@ -78,19 +78,19 @@ public class RuleExecution {
 
 
 
-    public Integer getStartEventID() {
+    public Long getStartEventID() {
         return startEventID;
     }
 
-    public void setStartEventID(Integer startEventID) {
+    public void setStartEventID(Long startEventID) {
         this.startEventID = startEventID;
     }
 
-    public Integer getStopEventID() {
+    public Long getStopEventID() {
         return stopEventID;
     }
 
-    public void setStopEventID(Integer stopEventID) {
+    public void setStopEventID(Long stopEventID) {
         this.stopEventID = stopEventID;
     }
